Working SAP Request
added sampleAction to Eclipse Menu and eventlistener.
added working SAP Backend Call from withing sampleMenu event
Request target is still invalid
